Reduced duplicated code on db-dribver dependent classes
