+++ src/Config/Option/AbstractNonSqliteDependentConfigOption.php
@@ -4,25 +4,10 @@ declare(strict_types=1);
 
 namespace Shlinkio\Shlink\Installer\Config\Option;
 
-use Shlinkio\Shlink\Installer\Util\PathCollection;
-
-abstract class AbstractNonSqliteDependentConfigOption implements
-    ConfigOptionInterface,
-    DependentConfigOptionInterface
+abstract class AbstractNonSqliteDependentConfigOption extends AbstractDriverDependentConfigOption
 {
-    public function getDependentOption(): string
+    protected function shouldBeAskedForDbDriver(string $dbDriver): bool
     {
-        return DatabaseDriverConfigOption::class;
-    }
-
-    public function shouldBeAsked(PathCollection $currentOptions, ?ConfigOptionInterface $dependantOption): bool
-    {
-        $currentOptionExists = $currentOptions->pathExists($this->getConfigPath());
-        if ($dependantOption === null) {
-            return ! $currentOptionExists;
-        }
-
-        $dbDriver = $currentOptions->getValueInPath($dependantOption->getConfigPath());
-        return $dbDriver !== DatabaseDriverConfigOption::SQLITE_DRIVER && ! $currentOptionExists;
+        return $dbDriver !== DatabaseDriverConfigOption::SQLITE_DRIVER;
     }
 }
