+++ src/Config/Option/DatabaseMySqlOptionsConfigOption.php
@@ -7,24 +7,11 @@ namespace Shlinkio\Shlink\Installer\Config\Option;
 use Shlinkio\Shlink\Installer\Util\PathCollection;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
-class DatabaseMySqlOptionsConfigOption implements
-    ConfigOptionInterface,
-    DependentConfigOptionInterface
+class DatabaseMySqlOptionsConfigOption extends AbstractDriverDependentConfigOption
 {
     public function getConfigPath(): array
     {
         return ['entity_manager', 'connection', 'driverOptions'];
-    }
-
-    public function shouldBeAsked(PathCollection $currentOptions, ?ConfigOptionInterface $dependantOption): bool
-    {
-        $currentOptionExists = $currentOptions->pathExists($this->getConfigPath());
-        if ($dependantOption === null) {
-            return ! $currentOptionExists;
-        }
-
-        $dbDriver = $currentOptions->getValueInPath($dependantOption->getConfigPath());
-        return $dbDriver === DatabaseDriverConfigOption::MYSQL_DRIVER && ! $currentOptionExists;
     }
 
     public function ask(
@@ -38,8 +25,8 @@ class DatabaseMySqlOptionsConfigOption implements
         ];
     }
 
-    public function getDependentOption(): string
+    protected function shouldBeAskedForDbDriver(string $dbDriver): bool
     {
-        return DatabaseDriverConfigOption::class;
+        return $dbDriver === DatabaseDriverConfigOption::MYSQL_DRIVER;
     }
 }
